WIP: fix one test, breaking 6 more [ci skip]
This is the code I presented in today's live stream.
The test for "Entity with trailing semicolon (1)" is now passing but
some tests are failing because of this fix.

I think it's something related to the parsing not giving back the chars
that were not consumed for the characted reference.